+++ app/assets/javascripts/orders_show.js
@@ -1,0 +1,1 @@
+//= require_tree ./orders_show
