+++ app/assets/javascripts/orders_show/orders_show.js
@@ -1,0 +1,3 @@
+$(document).ready(function(){
+  $("#cool").submit();
+});
