+++ app/models/order_item.rb
@@ -7,7 +7,7 @@ class OrderItem < ApplicationRecord
   def item_name
     options_arr = options.map{ |option|
       "#{option["name"]} (+#{option["price"]}元) "
-    }
-    "#{product["name"]} #{options_arr}"
+  }.join(" ")
+    "#{product["name"]}(#{product["price"]}元) #{options_arr}"
   end
 end
