+++ app/views/groups/buy.html.erb
@@ -55,7 +55,7 @@
         <div class="order-items media text-muted pt-3">
           <img data-src="holder.js/32x32?theme=thumb&bg=007bff&fg=007bff&size=1" alt="" class="mr-2 rounded">
           
-          <%= form_with url:"/groups/checkout", method: "get" do |form| %>
+          <%= form_with url: checkout_group_path, local: true do |form| %>
             <div id='order-item'>
             </div>
             <%= form.submit '結帳囉', :class => 'mt-3 btn btn-warning btn-sm' %>
