+++ config/initializers/assets.rb
@@ -20,3 +20,4 @@ Rails.application.config.assets.precompile += %w( devise.css )
 Rails.application.config.assets.precompile += %w( buy.css )
 Rails.application.config.assets.precompile += %w( buy.js )
 Rails.application.config.assets.precompile += %w( groups_link.js )
+Rails.application.config.assets.precompile += %w( orders_show.js )
